pyproject.toml
[tool.poetry]
name = "project-nebula"
version = "0.1.0"
description = ""
authors = ["Tilen Tratnjek <80243427+Tilchy@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.2"}
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
